Add selection state to selected contacts instead of removing them

+++ app/src/main/java/mega/privacy/android/app/lollipop/adapters/MegaContactsLollipopAdapter.java
@@ -37,7 +37,6 @@ import mega.privacy.android.app.lollipop.ManagerActivityLollipop;
 import mega.privacy.android.app.lollipop.listeners.UserAvatarListener;
 import mega.privacy.android.app.lollipop.managerSections.ContactsFragmentLollipop;
 import nz.mega.sdk.MegaApiAndroid;
-import nz.mega.sdk.MegaChatApi;
 import nz.mega.sdk.MegaChatApiAndroid;
 import nz.mega.sdk.MegaUser;
 
@@ -68,7 +67,6 @@ public class MegaContactsLollipopAdapter extends RecyclerView.Adapter<MegaContac
 	private SparseBooleanArray selectedItems;
 	private ContactsFragmentLollipop fragment;
 	private int adapterType;
-	private SparseBooleanArray selectedContacts;
 
 	DisplayMetrics outMetrics;
 
@@ -374,41 +372,29 @@ public class MegaContactsLollipopAdapter extends RecyclerView.Adapter<MegaContac
 
 		UserAvatarListener listener = new UserAvatarListener(context, holder);
 
-		File avatar = buildAvatarFile(context, holder.contactMail + ".jpg");
-		Bitmap bitmap = null;
-		if (isFileAvailable(avatar)){
-			if (avatar.length() > 0){
-				BitmapFactory.Options bOpts = new BitmapFactory.Options();
-				bOpts.inPurgeable = true;
-				bOpts.inInputShareable = true;
-				bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-				if (bitmap == null) {
-					avatar.delete();
-                    megaApi.getUserAvatar(contact.getMegaUser(),buildAvatarFile(context,contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(),listener);
-                }
-				else{
-					holder.imageView.setImageBitmap(bitmap);
+		if (contact.isSelected()) {
+			holder.itemLayout.setBackgroundColor(ContextCompat.getColor(context, R.color.new_multiselect_color));
+		} else {
+			File avatar = buildAvatarFile(context, holder.contactMail + ".jpg");
+			Bitmap bitmap = null;
+			if (isFileAvailable(avatar)) {
+				if (avatar.length() > 0) {
+					BitmapFactory.Options bOpts = new BitmapFactory.Options();
+					bOpts.inPurgeable = true;
+					bOpts.inInputShareable = true;
+					bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+					if (bitmap == null) {
+						avatar.delete();
+						megaApi.getUserAvatar(contact.getMegaUser(), buildAvatarFile(context, contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(), listener);
+					} else {
+						holder.imageView.setImageBitmap(bitmap);
+					}
+				} else {
+					megaApi.getUserAvatar(contact.getMegaUser(), buildAvatarFile(context, contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(), listener);
 				}
+			} else {
+				megaApi.getUserAvatar(contact.getMegaUser(), buildAvatarFile(context, contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(), listener);
 			}
-			else{
-                megaApi.getUserAvatar(contact.getMegaUser(),buildAvatarFile(context,contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(),listener);
-			}
-		}
-		else{
-            megaApi.getUserAvatar(contact.getMegaUser(),buildAvatarFile(context,contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(),listener);
-		}
-
-		if (selectedContacts != null) {
-			for (int i = 0; i < selectedContacts.size(); i++) {
-				if (selectedContacts.get(position) == true) {
-					holder.itemLayout.setBackgroundColor(ContextCompat.getColor(context, R.color.new_file_list_selected_row));
-				}
-				else{
-					holder.itemLayout.setBackgroundColor(Color.WHITE);
-				}
-			}
-		}
-		else{
 			holder.itemLayout.setBackgroundColor(Color.WHITE);
 		}
 	}
@@ -461,20 +447,6 @@ public class MegaContactsLollipopAdapter extends RecyclerView.Adapter<MegaContac
 		}
 		else{
             megaApi.getUserAvatar(contact.getMegaUser(),buildAvatarFile(context,contact.getMegaUser().getEmail() + ".jpg").getAbsolutePath(),listener);
-		}
-
-		if (selectedContacts != null) {
-			for (int i = 0; i < selectedContacts.size(); i++) {
-				if (selectedContacts.get(position) == true) {
-					holder.itemLayout.setBackgroundColor(ContextCompat.getColor(context, R.color.new_file_list_selected_row));
-				}
-				else{
-					holder.itemLayout.setBackgroundColor(Color.WHITE);
-				}
-			}
-		}
-		else{
-			holder.itemLayout.setBackgroundColor(Color.WHITE);
 		}
 	}
 	
@@ -947,11 +919,6 @@ public class MegaContactsLollipopAdapter extends RecyclerView.Adapter<MegaContac
 		fragment.itemClick(currentPosition);
 
 		return true;
-	}
-
-	public void setSelectedContacts(SparseBooleanArray selectedContacts){
-		this.selectedContacts = selectedContacts;
-		notifyDataSetChanged();
 	}
 
 	public MegaUser getDocumentAt(int position) {
